WIP dynamic IR repeats with callback function
The initial repeat count is still taken from the IR send function.
The callback is called during the repeat sequence in the IR specific
protocol sending function to check if the repeat count needs to
be extended.
Note: the initial repeat value cannot be shortened, only prolonged!

The specific logic on how the repeat count is prolonged is up to the
client. It can either count the number of callbacks to act on specific
counts, or set a timeout when it signals to no longer repeat the IR code.